app/backend/requirements/base.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
colorlog==6.7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-enumfields==2.1.1
django-extensions==3.2.1
django-filter==22.1
django-js-asset==2.0.0
django-money==3.0.0
django-otp==1.1.4
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
flower==1.2.0
frozenlist==1.3.3
hedera-sdk-py==2.17.3
humanize==4.6.0
idna==3.4
inflection==0.5.1
install==1.3.5
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
JPype1==1.4.0
jsonapi-client==0.9.9
jsonschema==4.17.3
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.4
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
py-moneyed==2.0
pycparser==2.21
pyjnius==1.4.2
PyJWT==2.6.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.6
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
yarl==1.8.2
